edits on lab 11 ex 4

+++ src/Read.java
@@ -2,15 +2,13 @@ import java.util.InputMismatchException;
 import java.util.Scanner;
 
 public class Read {
-    //   This scans for numerical input and prompts you to repeat till it gets numerical value ONLY!
-
-    public double getDouble() {
+    public int getInt() {
         Scanner MyNewScanner = new Scanner(System.in);
-        double x;
+        int x;
         do {
             try {
                 System.out.println("Enter a number");
-                x = MyNewScanner.nextDouble();
+                x = MyNewScanner.nextInt();
                 return x;
             } catch (InputMismatchException e) {
                 System.err.println("This number '" + MyNewScanner.nextLine() + "is wrong. Do it again!");
